NetCDF file to Excel
Basically the problem was we needed to extract some data from some NetCDF file for using in Delft 3D model which neede specific pattern of input data of wind and wave. But the file contained many unnecessary data that we did not need. Also, the file contained data of 16 locations. But we needed data of just 8 locations.